requirements.txt
bio~=1.3.3
stringdb~=0.1.5
mysql-connector-python~=8.0.28
matplotlib~=3.5.2
numpy~=1.22.3
scipy~=1.8.0
pandas~=1.4.2
sklearn~=0.0
scikit-learn~=1.0.2